+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/World.java
@@ -20,7 +20,7 @@ public final class World extends DomainModel {
 
 	public World() {
 		super(DOMAIN);
-		setStrictForeignKeys(false);
+		validateForeignKeys(false);
 		add(countryCityV(), city(), country(),
 				countrylanguage());
 	}